requirements.txt
beautifulsoup4==4.8.2
html5lib==1.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.21
numpy==1.18.5
pipwin==0.4.9
PyPrind==2.11.2
PyQt5==5.15.1
PyYAML==5.3.1
requests==2.23.0
selenium==3.141.0
urllib3==1.25.8

